+++ src/data/challenges/get-challenges.ts
@@ -29,9 +29,6 @@ export const getChallenges = async () => {
           where: {
             userId: session.user.id
           },
-          select: {
-            passed: true
-          }
         }
       }
     });
